Add several new placeholders
This commit adds several new placeholders:

* 'trainer_team_#_hiddenpower' - The type of the pokemon's  Hidden Power
move. If the pokemon doesn't know this move, then this placeholder
returns the type that it would be if they were to learn it.
* 'trainer_team_#_ivtotalsum' - The sum of the pokemon's IV stats
* 'trainer_team_#_evtotalsum' - The sum of the pokemon's EV stats
* 'trainer_team_#_ivtotalpercent' - The sum of the pokemon's IV stats as
a percentage of the maximum possible IV sum
* 'trainer_team_#_evtotalpercent' - The sum of the pokemon's EV stats as
a percentage of the maximum possible EV sum
* 'trainer_team_#_texturelocation' - The ResourceLocation of the
pokemon's sprite texture.
* 'trainer_team_#_customtexture' - The ResourcELocation of the pokemon's
custom texture, or 'N/A' if the pokemon has no custom texture
* 'trainer_team_#_form' - The form number of the pokemon

Additionally, this commit adds integration with the 'entity-particles'
plugin. If the 'entity-particles' is installed, the following
placeholder will be enabled:

* 'trainer_team_#_aura' - The ID of the 'entity-particles' effect
currently possesed by the pokemon.

If the pokemon has no 'entity-particles' effect, or 'entity-particles'
is not installed, the placeholder returns 'N/A'